UIMARCAUTH-170: Tweak the search and view authority components
