--- src/components/SearchResultsList/index.js
@@ -1,1 +1,0 @@
-export { default as SearchResultsList } from './SearchResultsList';
