--- src/components/SearchResultsList/useHighlightEditedRecord/index.js
@@ -1,1 +1,0 @@
-export * from './useHighlightEditedRecord';
