+++ src/components/index.js
@@ -1,3 +1,2 @@
-export * from './SearchResultsList';
 export * from './MarcAuthoritiesAppContext';
 export * from './KeyShortCutsWrapper';
